backend/besm6: spell the 48-bit machine word uint64_t, not unsigned long long
Every use of `unsigned long long` in the backend models the BESM-6's 48-bit
word: the 41/48-bit masks, the loops that pack six bytes into a word, and the
floating-point encoder that lays a 7-bit exponent at bit 41.  That is an
exact-width contract, so name it with an exact-width type.  static.c already
mixed int64_t/uint8_t casts into unsigned long long code, and the rest of the
tree (tac/, ast/, semantic/) settled on <stdint.h> plus <inttypes.h> PRI*
macros long ago.

Also fixes a latent portability bug: the five `%llo` conversions become
PRIo64.  On LP64 Linux uint64_t is `unsigned long`, where `%llo` would trip
-Wformat under -Werror.

Emitted assembly is unchanged.  Verified by building HEAD in a scratch
worktree and byte-comparing both dialects' output for a program covering all
eight static-init widths, strings, multi-dimensional char arrays, packed
struct members, every Tac_Const kind, and the -0.5/-1.0 renormalization edges
of unix_real_word: Madlen and Unix output are identical, with the TAC as a
control.

CLAUDE.md: genbesm defaults to the Unix (b6as) dialect, not Madlen, and picks
the output extension from the dialect.  The pipeline diagram and the
hand-built-Dubna-job recipe both implied Madlen was the default.

+++ backend/besm6/besm.h
@@ -2,6 +2,7 @@
 #define BESM_H
 
 #include <stdbool.h>
+#include <stdint.h>
 #include <stdio.h>
 
 #ifdef __cplusplus
@@ -107,18 +108,18 @@ typedef enum {
 struct Besm_Instr {
     struct Besm_Instr *next;
     Besm_InstrKind kind;
-    unsigned reg;               // index-register 0..15, or displacement (BESM_DATA_Z00)
-    int addr;                   // offset, or integer data value (BESM_DATA_INT/BSS/EQU)
-    char *name;                 // symbolic name, optional (heap-owned)
-    char *label;                // Madlen label for a data word whose `name` is already an
-                                // operand (currently only BESM_DATA_Z00), optional (heap-owned)
-    struct Tac_Const *konst;    // scalar nonzero constant operand (=N / #N / =Ю'…'), optional
-                                // (heap-owned); when set, the operand is this literal, not
-                                // (name, addr).  The per-dialect emitter formats it.  A zero
-                                // constant is never attached: the instruction is left with no
-                                // operand and reads memory word 0 (see emit.c attach_const).
-    unsigned long long log_val; // BESM_DATA_LOG: 48-bit logical constant
-    double real_val;            // BESM_DATA_REAL
+    unsigned reg;            // index-register 0..15, or displacement (BESM_DATA_Z00)
+    int addr;                // offset, or integer data value (BESM_DATA_INT/BSS/EQU)
+    char *name;              // symbolic name, optional (heap-owned)
+    char *label;             // Madlen label for a data word whose `name` is already an
+                             // operand (currently only BESM_DATA_Z00), optional (heap-owned)
+    struct Tac_Const *konst; // scalar nonzero constant operand (=N / #N / =Ю'…'), optional
+                             // (heap-owned); when set, the operand is this literal, not
+                             // (name, addr).  The per-dialect emitter formats it.  A zero
+                             // constant is never attached: the instruction is left with no
+                             // operand and reads memory word 0 (see emit.c attach_const).
+    uint64_t log_val;        // BESM_DATA_LOG: 48-bit logical constant
+    double real_val;         // BESM_DATA_REAL
     // TODO: star plus offset
     // TODO: literal address with value (int, uns, real)
 };
